+++ sst.config.ts
@@ -22,7 +22,7 @@ export default $config({
     };
   },
   async run() {
-    const hono = new sst.aws.Function("Api", {
+    const api = new sst.aws.Function("Api", {
       architecture: "arm64",
       description: "Hono API for LTA Cars Datasets",
       environment: {
@@ -39,12 +39,55 @@ export default $config({
     new sst.aws.Router("LTACarsDataset", {
       domain: DOMAIN[$app.stage],
       routes: {
-        "/*": hono.url,
+        "/*": api.url,
       },
     });
 
+    // TODO: To remove the cron scheduler after migrating to the python
+    if ($app.stage === "production") {
+      new sst.aws.Cron("UpdateCarsJob", {
+        schedule: "cron(0/60 0-10 ? * MON-FRI *)",
+        job: {
+          handler: "src/updater.cars",
+          environment: {
+            MONGODB_URI: process.env.MONGODB_URI,
+          },
+        },
+      });
+
+      new sst.aws.Cron("UpdateCOEJob", {
+        schedule: "cron(0/60 0-10 ? * MON-FRI *)",
+        job: {
+          handler: "src/updater.coe",
+          environment: {
+            MONGODB_URI: process.env.MONGODB_URI,
+          },
+        },
+      });
+
+      new sst.aws.Cron("UpdateCOEFirstBiddingJob", {
+        schedule: "cron(0/10 8-10 ? * 4#1 *)",
+        job: {
+          handler: "src/updater.coe",
+          environment: {
+            MONGODB_URI: process.env.MONGODB_URI,
+          },
+        },
+      });
+
+      new sst.aws.Cron("UpdateCOESecondBiddingJob", {
+        schedule: "cron(0/10 8-10 ? * 4#3 *)",
+        job: {
+          handler: "src/updater.coe",
+          environment: {
+            MONGODB_URI: process.env.MONGODB_URI,
+          },
+        },
+      });
+    }
+
     return {
-      api: hono.url,
+      api: api.url,
     };
   },
 });
